backend/requirements.txt
schema==0.6.6
django==1.11.5
djangorestframework==3.6.4
gunicorn==19.7.1